feat: implement login page
Implements the login page for the KSPM Payment Center app.

Adds a new page named 'Login' which contains a login form.
The login form consists of a KSPM-UTY logo, two text fields for NIM and password, a login button, and a forgot password link.
Also includes a splash screen with a KSPM-UTY logo which redirects to the login page after 3 seconds.